Make ActionListener.notifyOnce a little cheaper
This is used on the rather hot path now due to #97301, lets apply the optimization
of saving one level of indirection here as well to make GC etc. a little cheaper on these.
